extracted api routes to handlers.rs
